6.12.9_2024.01.03  extends and privacizes Variables.Cell
Cell now maintains all methods relating to unit cells and their axes.

It replaces the various Variables.s.... and Variables.p.... arrays and
much simplifies access to useful methods such as toTempCartesian().

It incorporates cellInfo and axesInfo from IsoDistortApp.

Many of the methods of IsoDistortApp are greatly simplified, as the
parentCell and childCell instances take care of all the low-level
lattice operations.